欢迎语附件相关代码优化提交

+++ src/main/java/cn/iyque/controller/IYqueUserCodeController.java
@@ -125,7 +125,7 @@ public class IYqueUserCodeController {
      * @return
      */
     @GetMapping("/distributeUserCode/{id}")
-    public ResponseResult distributeUserCode(@PathVariable("id") Integer id){
+    public ResponseResult distributeUserCode(@PathVariable("id") Long id){
 
         try {
             iYqueUserCodeService.distributeUserCode(id);
